Function that simulates the mean window period, mean duration of recent infection, and false recent rate (FRR) based on external study data.

The external study data for mean window and MDRI is available here from Duong et al. 2015: https://doi.org/10.1371/journal.pone.0114947.s001

assay.properties.nsim(
  n_sims,
  phi.func,
  bigT,
  tau,
  ext_FRR = FALSE,
  ext_df = NULL,
  max_FRR = NULL,
  last_point = FALSE
)

Arguments

n_sims
Number of simulations

phi.func
A test-recent positive function of t

bigT
The time cutoff value designating true recent versus false recent

tau
The maximum duration of infection where a subject could have a false-positive for recent infection

ext_FRR
Whether or not to get FRR from the Duong et al. 2015 study or to calculate it from a simpler, separate study, passed in ext_df

ext_df
A dataset with column "duration" and column for binary "recent" indicator

max_FRR
The maximum duration allowed

last_point
Integrate the mean window period to the last observed duration in the dataset, rather than tau

Value

A list of estimated mean window period \(\mu\) and its variance, estimated MDRI \(\Omega_{T^*}\) and its variance, and estimated FRR \(\beta_{T^*}\) and its variance.

Examples

set.seed(0)
assay.properties.nsim(n_sims=1, phi.func=function(t) 1 - pgamma(t, 1, 1.5),
                     bigT=2, tau=12)

#> Error in geese(gap ~ samp + samp.5 + samp.5 * samp, id = id.key, data = df,     family = poisson(link = "log"), corstr = "exchangeable"): could not find function "geese"