Function that simulates the mean window period, mean duration of recent infection, and false recent rate (FRR) based on external study data.

Function that simulates the mean window period, mean duration of recent infection, and false recent rate (FRR) based on external study data.

assay.properties.sim(phi.func, bigT, tau, frr_mixture = NULL)

Arguments

phi.func
A test-recent positive function of t

bigT
The time cutoff value designating true recent versus false recent

tau
The maximum duration of infection where a subject could have a false-positive for recent infection

frr_mixture
Vector of two values. To simulate data for FRR, we need a sample of long infected subjects. If frr_mixture is NULL, then the sample of infection times is \(Uniform(2, 12)\). If frr_mixture is not NULL, then the sample is an equal mixture of \(Uniform(2, 12)\) and \(Uniform(x[1], x[2])\) where \(x\) is frr_mixture.

Value

A list of estimated mean window period \(\mu\) and its variance, estimated MDRI \(\Omega_{T^*}\) and its variance, and estimated FRR \(\beta_{T^*}\) and its variance.

Examples

set.seed(0)
assay.properties.sim(phi.func=function(t) 1 - pgamma(t, 1, 1.5),
                     bigT=2, tau=12)

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> $mu_est
#> [1] 0.647143
#> 
#> $mu_var
#> [1] 0.0005193913
#> 
#> $omega_est
#> [1] 0.6335523
#> 
#> $omega_var
#> [1] 0.0004514846
#> 
#> $beta_est
#> [1] 0.004322767
#> 
#> $beta_var
#> [1] 3.100922e-06
#> 

set.seed(0)
assay.properties.sim(phi.func=function(t) 1 - pgamma(t, 1, 1.5),
                     bigT=2, tau=12, frr_mixture=c(5, 8))

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> $mu_est
#> [1] 0.647143
#> 
#> $mu_var
#> [1] 0.0005193913
#> 
#> $omega_est
#> [1] 0.6335523
#> 
#> $omega_var
#> [1] 0.0004514846
#> 
#> $beta_est
#> [1] 0.002881844
#> 
#> $beta_var
#> [1] 2.070273e-06
#>