Create simulations of trials based on a past infection time function and a prevalence and baseline incidence (time 0).

Create simulations of trials based on a past infection time function and a prevalence and baseline incidence (time 0).

generate.data(
  n_sims,
  n,
  infection.function,
  baseline_incidence,
  prevalence,
  rho,
  phi.func = NULL,
  times = c(0),
  summarize = TRUE
)

Arguments

n_sims
Number of simulations

n
Number of subjects screened

infection.function
Function that simulates the infection time

baseline_incidence
Baseline incidence value (time 0)

prevalence
Constant prevalence value

rho
A parameter to the infection.function for how quickly incidence changes

phi.func
Optional test positive function for recency assay. If you pass summarize=TRUE, then you must provide a test positive function.

times
A vector of times at which to enroll (e.g. c(0, 1, 2, 3))

summarize
Whether to return a summary of the dataset or unit-record data

Value

If summarize=TRUE, a list of vectors (or matrices if length(times > 1) for summary screening quantities across simulations (and potentially times if length(times) > 1):

n: number screened

times: times screened

n_p: number of positives

n_n: number of negatives

n_r: number of recents

If summarize=FALSE, then returns a list of data frames by simulation that have unit-record data with infection times with column names:

time: time screened

pos: indicator for positive

itime: infection time (reference time 0)

probs: recency test positive probability based on infection time

rpos: indicator for recency test positive, if a phi function is passed

Examples

set.seed(1)
generate.data(n_sims=3, n=2, infection.function=con.infections,
              baseline_incidence=0.05, prevalence=0.5, rho=NA,
              phi.func=function(t) 1 - pgamma(t, 1, 2),
              times=c(0, 1), summarize=FALSE)

#> Error in simulate.recent(data, infection.function, phi.func, baseline_incidence,     prevalence, rho, summarize = summarize): object 'con.infections' not found

generate.data(n_sims=3, n=100, infection.function=con.infections,
              baseline_incidence=0.05, prevalence=0.3, rho=NA,
              phi.func=function(t) 1 - pgamma(t, 1, 2),
              times=c(0, 1), summarize=TRUE)

#> Error in simulate.recent(data, infection.function, phi.func, baseline_incidence,     prevalence, rho, summarize = summarize): object 'con.infections' not found

generate.data(n_sims=3, n=100, infection.function=con.infections,
              baseline_incidence=0.05, prevalence=0.3, rho=NA,
              phi.func=function(t) 1 - pgamma(t, 1, 2),
              times=c(0), summarize=TRUE)

#> Error in simulate.recent(data, infection.function, phi.func, baseline_incidence,     prevalence, rho, summarize = summarize): object 'con.infections' not found