Get enhanced estimate and variance (Bannick and Gao 2023+)

Get enhanced estimate and variance (Bannick and Gao 2023+)

get.adjusted.pt(
  n_p,
  n,
  ptdf,
  beta,
  beta_var,
  big_T,
  phidat,
  use_geese,
  formula,
  family,
  plot_phi = TRUE,
  return_all = FALSE,
  ...
)

Arguments

n_p

Number of positives

n

Number of observations

ptdf

Prior testing data frame. Only include those who are positive.

beta

False recency rate (FRR)

beta_var

Variance of the estimator for FRR (or 0 if FRR known)

big_T

The \(T^*\) in the equation above

phidat

Data to fit a modfunc model to. Must have column names ri for recency indicator and ui for infection duration. Can have more columns, e.g., if an id is needed for geese.

use_geese

Indicator to fit a gee model using geese(), or a glm model with glm()

formula

Formula for fitting the model to phidat. Formula argument must take in only ui as the newdata. For example, do not create a ui^2 variable. Use poly(ui, ...) function to fit polynomial terms.

family

Family argument for glm or gee

plot_phi

Whether to plot the estimated phi function

Value

Returns a list of the estimate and the variance.

Examples

set.seed(101)

# Get HIV, recency, and prior testing data
e.func <- function(e) infections.con(e, t=0, p=0.29, lambda=0.032)
phi.func <- function(t) 1-pgamma(t, shape=1, rate=2)
sim <- sim.screening.generator(prevalence=0.29, e.func=e.func, phi.func=phi.func)
df <- sim(1000)
sim.pt <- sim.pt.generator(ptest.dist=function(u) runif(1, 0, 4),
                           ptest.prob=function(u) 0.1)
ptdf <- sim.pt(df[df$di == 1,])

# Get dataset for estimating \Omega
phidat <- get.assay.df(assays=c("LAg-Sedia"),
                       algorithm=function(l) ifelse(l <= 1.5, 1, 0))

get.adjusted.pt(
  n_p=nrow(ptdf),
  n=nrow(df),
  ptdf=ptdf,
  beta=0,
  beta_var=0,
  big_T=1,
  phidat=phidat,
  use_geese=TRUE,
  formula="ri ~ poly(ui, degree=3, raw=T)",
  family=binomial(link="logit")
)
#> Error in estimate.phi(phidat = phidat, maxT = maxT, bigT = bigT, dt = dt,     formula = formula, family = family, use_geese = use_geese,     plot_phi = FALSE, return_all = TRUE, ...): object 'geese' not found