Modifies the prior testing data for recall bias.

Modifies the prior testing data for recall bias.

modifyPriorTests(
  t_noise = function(t) t,
  t_range = NULL,
  d_misrep = 0,
  p_misrep = 0
)

Arguments

t_noise

Function to add noise to a single prior testing time. Must be a function of t, the prior testing time.

t_range

Range of times in which to keep prior testing times.

d_misrep

Probability of given a positive prior test, individual reports negative prior test.

p_misrep

Probability of not reporting a prior test result, among those with positive prior test.

Examples

e.func <- function(e) infections.lin(e, t=0, p=0.29, lambda=0.032, rho=0.07)
params <- get.gamma.params(window=200/365.25, shadow=191/365.25)
#> Error in get.gamma.params(window = 200/365.25, shadow = 191/365.25): could not find function "get.gamma.params"
phi.func <- function(t) 1-pgamma(t, shape = params[1], rate = params[2])
sim <- simCrossSect(prevalence=0.29, e.func=e.func, phi.func=phi.func)
df <- sim(100)
#> Error in infections.lin(e, t = 0, p = 0.29, lambda = 0.032, rho = 0.07): could not find function "infections.lin"
sim.pt <- simPriorTests(ptest.dist=function(u) runif(1, 0, 4),
                         ptest.prob=function(u) 0.5)
pt.df <- sim.pt(df[df$di == 1,])
#> Error in df$di: object of type 'closure' is not subsettable
t_noise <- function(t) max(0, t + rnorm(n=1, sd=0.5))
modifypt <- modifyPriorTests(t_noise=t_noise, d_misrep=1, p_misrep=0, t_range=c(1, 2))
pt.df2 <- modifypt(pt.df)
#> Error in lapply(X = X, FUN = FUN, ...): object 'pt.df' not found